requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
black==25.1.0
bson==0.5.10
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
distro==1.9.0
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
griffe==1.7.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
isort==6.0.1
jiter==0.9.0
mcp==1.8.0
mypy_extensions==1.1.0
openai==1.78.0
openai-agents==0.0.14
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
sse-starlette==2.3.4
starlette==0.46.2
tqdm==4.67.1
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
